requirements.locals.txt
Flask==0.10.1
Flask-Gravatar==0.2.3
Flask-SQLAlchemy==0.15
Jinja2==2.6
PyYAML==3.11
SQLAlchemy==0.7.2
Werkzeug==0.8.3
feedformatter==0.4
feedparser==5.1.2
httplib2==0.7.4
wsgiref==0.1.2
flake8
